Changed return types

+++ lwlog/src/details/os/os.h
@@ -12,24 +12,24 @@
 namespace lwlog::details::os
 {
 	template<typename ThreadIdPolicy>
-	static std::size_t get_thread_id();
+	static std::uint64_t get_thread_id();
 
 	template<typename ProcessIdPolicy>
-	static std::size_t get_process_id();
+	static std::uint64_t get_process_id();
 
 	struct execution_context_base
 	{
 		virtual ~execution_context_base() = default;
 
-		virtual std::size_t thread_id() const = 0;
-		virtual std::size_t process_id() const = 0;
+		virtual std::uint64_t thread_id() const = 0;
+		virtual std::uint64_t process_id() const = 0;
 	};
 
 	template<typename ThreadIdPolicy, typename ProcessIdPolicy>
 	struct execution_context : public execution_context_base
 	{
-		std::size_t thread_id() const override;
-		std::size_t process_id() const override;
+		std::uint64_t thread_id() const override;
+		std::uint64_t process_id() const override;
 	};
 
 	static bool are_ansi_colors_enabled();
